+++ Cargo.toml
@@ -48,3 +48,4 @@ serde_json = "1.0"
 toml = "0.8"
 futures-lite = "2.3"
 thiserror = "1.0"
+tokio = { version = "1.37", default-features = false }
