--- support/device/TODO
@@ -1,3 +1,0 @@
-
-README:
-- add `libudev-sys` to linux deps
