+++ support/device/src/usb/discover.rs
@@ -11,29 +11,31 @@ use crate::VENDOR_ID;
 
 use super::Device;
 
+type Result<T, E = Error> = std::result::Result<T, E>;
+
 
 /// Enumerate all Playdate- devices.
 #[cfg_attr(feature = "tracing", tracing::instrument)]
-pub fn devices() -> Result<impl Iterator<Item = Device>, Error> {
+pub fn devices() -> Result<impl Iterator<Item = Device>> {
 	Ok(nusb::list_devices()?.filter(|d| d.vendor_id() == VENDOR_ID)
 	                        .map(|info| Device::new(info)))
 }
 
 /// Search Playdate- devices that in data (serial/modem/telnet) mode.
 #[cfg_attr(feature = "tracing", tracing::instrument)]
-pub fn devices_data() -> Result<impl Iterator<Item = Device>, Error> {
+pub fn devices_data() -> Result<impl Iterator<Item = Device>> {
 	devices().map(|iter| iter.filter(|d| d.info.product_id() == PRODUCT_ID_DATA))
 }
 
 /// Search Playdate- devices that in storage (data-disk) mode.
 #[cfg_attr(feature = "tracing", tracing::instrument)]
-pub fn devices_storage() -> Result<impl Iterator<Item = Device>, Error> {
+pub fn devices_storage() -> Result<impl Iterator<Item = Device>> {
 	devices().map(|iter| iter.filter(|d| d.info.product_id() == PRODUCT_ID_STORAGE))
 }
 
 /// Search exact one device with same serial number.
 #[cfg_attr(feature = "tracing", tracing::instrument)]
-pub fn device(sn: &Sn) -> Result<Device, Error> {
+pub fn device(sn: &Sn) -> Result<Device> {
 	devices()?.find(|d| d.info.serial_number().filter(|s| sn.eq(s)).is_some())
 	          .ok_or_else(|| Error::not_found())
 }
@@ -41,7 +43,7 @@ pub fn device(sn: &Sn) -> Result<Device, Error> {
 /// Search devices with same serial number,
 /// or __any__ Playdate- device if `sn` is `None`.
 #[cfg_attr(feature = "tracing", tracing::instrument)]
-pub fn devices_with(sn: Option<Sn>) -> Result<impl Iterator<Item = Device>, Error> {
+pub fn devices_with(sn: Option<Sn>) -> Result<impl Iterator<Item = Device>> {
 	Ok(devices()?.filter(move |dev| {
 		             if let Some(sn) = sn.as_ref() {
 			             dev.info().serial_number().filter(|s| sn.eq(s)).is_some()
@@ -54,7 +56,7 @@ pub fn devices_with(sn: Option<Sn>) -> Result<impl Iterator<Item = Device>, Erro
 /// Search devices with same serial number in data mode,
 /// or __any__ Playdate- device if `sn` is `None`.
 #[cfg_attr(feature = "tracing", tracing::instrument)]
-pub fn devices_data_with(sn: Option<Sn>) -> Result<impl Iterator<Item = Device>, Error> {
+pub fn devices_data_with(sn: Option<Sn>) -> Result<impl Iterator<Item = Device>> {
 	Ok(devices_data()?.filter(move |dev| {
 		                  if let Some(sn) = sn.as_ref() {
 			                  dev.info().serial_number().filter(|s| sn.eq(s)).is_some()
@@ -67,7 +69,7 @@ pub fn devices_data_with(sn: Option<Sn>) -> Result<impl Iterator<Item = Device>,
 
 #[cfg(feature = "futures")]
 #[cfg_attr(feature = "tracing", tracing::instrument)]
-pub async fn devices_data_for(query: query::Query) -> Result<Vec<Device>, Error> {
+pub async fn devices_data_for(query: query::Query) -> Result<Vec<Device>> {
 	use query::Value as Query;
 	use serial::dev_with_port;
 
@@ -109,7 +111,7 @@ pub async fn devices_data_for(query: query::Query) -> Result<Vec<Device>, Error>
 				Ok(None) => dev_with_port(port_pref).await,
 				Err(err) => {
 					dev_with_port(port_pref).await
-					                        .map_err(|err2| Error::chain(err, [err2]))
+					                        .map_err(|err2| Error::chain(err2, [err]))
 				},
 			}
 		}
@@ -131,12 +133,9 @@ pub async fn devices_data_for(query: query::Query) -> Result<Vec<Device>, Error>
 
 #[cfg(feature = "futures")]
 #[cfg_attr(feature = "tracing", tracing::instrument(skip(f)))]
-pub async fn for_each_data_interface<F, Fut, T>(query: query::Query,
-                                                mut f: F)
-                                                -> Result<impl Stream<Item = T>, Error>
-	where Fut: std::future::Future<Output = T>, /* + Unpin */
-	      F: FnMut(interface::Interface) -> Fut
-{
+pub async fn for_each_data_interface<F, Fut, T>(query: query::Query, mut f: F) -> Result<impl Stream<Item = T>>
+	where Fut: std::future::Future<Output = T>,
+	      F: FnMut(interface::Interface) -> Fut {
 	use query::Value as Query;
 	use serial::unknown_serial_port_info;
