+++ support/tool/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "playdate-tool"
-version = "0.1.4"
+version = "0.2.0"
 readme = "README.md"
 description = "Tool for interaction with Playdate device and sim."
 keywords = ["playdate", "usb", "utility"]
@@ -15,32 +15,52 @@ repository.workspace = true
 [[bin]]
 path = "src/main.rs"
 name = "pdtool"
-required-features = ["cli"]
 
 
 [dependencies]
-regex.workspace = true
-log.workspace = true
-env_logger = { workspace = true, optional = true }
-thiserror = "1.0"
+# RT, async:
+tokio = { version = "1.36", features = ["full", "rt-multi-thread"] }
+futures = { version = "0.3" }
+futures-lite.workspace = true
 
+# fmt:
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
-plist = "1.6"
 
-rusb = { version = "0.9", optional = true }
-usb-ids = { version = "1.2023.0", optional = true }
+# CLI:
+log.workspace = true
+env_logger.workspace = true
+thiserror.workspace = true
+miette = { version = "7.2", features = ["fancy"] }
 
-# used assert::env-resolver and sdk
-build = { workspace = true, default-features = false }
+# tracing:
+tracing = { version = "0.1", optional = true }
+tracing-subscriber = { version = "0.3", optional = true }
+console-subscriber = { version = "0.2", features = [
+	"env-filter",
+], optional = true }
 
 [dependencies.clap]
 features = ["std", "env", "derive", "help", "usage", "color"]
 workspace = true
-optional = true
+
+# PD:
+[dependencies.device]
+features = ["async", "tokio", "clap", "tokio-serial"]
+workspace = true
+
+[dependencies.simulator]
+features = ["tokio"]
+workspace = true
 
 
 [features]
-default = ["usb"]
-cli = ["clap", "env_logger"]
-usb = ["rusb", "usb-ids"]
+tracing = [
+	"dep:tracing",
+	"tracing-subscriber",
+	"device/tracing",
+	"simulator/tracing",
+]
+
+# Useful with [tokio-console](https://tokio.rs/tokio/topics/tracing-next-steps)
+tokio-tracing = ["tracing", "tokio/tracing", "console-subscriber", "tracing"]
