--- support/tool2/Cargo.toml
@@ -1,66 +1,0 @@
-[package]
-name = "playdate-tool2"
-version = "0.2.0-alpha1"
-readme = "README.md"
-description = "Tool for interaction with Playdate device and sim."
-keywords = ["playdate", "usb", "utility"]
-categories = ["development-tools", "hardware-support"]
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-
-[[bin]]
-path = "src/main.rs"
-name = "pdtool"
-# required-features = ["tracing"]
-
-
-[dependencies]
-# RT, async:
-tokio = { version = "1.36", features = ["full", "rt-multi-thread"] }
-futures = { version = "0.3" }
-futures-lite.workspace = true
-
-serde = { workspace = true, features = ["derive"] }
-serde_json.workspace = true
-
-# CLI:
-log.workspace = true
-env_logger.workspace = true
-thiserror.workspace = true
-miette = { version = "7.2", features = ["fancy"] }
-
-tracing = { version = "0.1", optional = true }
-tracing-subscriber = { version = "0.3", optional = true }
-console-subscriber = { version = "0.2", features = [
-	"env-filter",
-], optional = true }
-
-[dependencies.clap]
-features = ["std", "env", "derive", "help", "usage", "color"]
-workspace = true
-
-# PD:
-[dependencies.device]
-features = ["async", "tokio", "clap", "tokio-serial"]
-workspace = true
-
-[dependencies.simulator]
-features = ["tokio"]
-workspace = true
-
-
-[features]
-default = ["tracing"]
-tracing = [
-	"dep:tracing",
-	"tracing-subscriber",
-	"device/tracing",
-	"simulator/tracing",
-]
-
-# use with tokio-console, https://tokio.rs/tokio/topics/tracing-next-steps
-tokio-tracing = ["tracing", "tokio/tracing", "console-subscriber", "tracing"]
